pyproject.toml
[tool.poetry]
name = "pfeffel"
version = "0.1.0"
description = ""
authors = ["The Pfeffel team "]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.20.1"
pandas = "^1.4.2"
xlrd = "^2.0.1"
tqdm = "^4.64.0"
pipe = "2.2"
requests = "^2.28.0"
PyYAML = "^6.0"
jupyterlab = "^3.4.3"
folium = "^0.7.0"
seaborn = "^0.10.0"
matplotlib = "^3.5.2"
ipywidgets = "^7.7.0"
h3 = "^3.7.4"
scipy = "^1.8.1"
Cartopy = "^0.22"
networkx = "^2.8.4"
python-louvain = "^0.16"
movingpandas = "^0.9rc3"
stonesoup = "^0.1b9"
contextily = "^1.2.0"
fiona = "^1.9"


[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
pre-commit = "^2.19.0"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
